config discovery: keep python bridge JSON-only
