+++ pkg/collector/python/BUILD.bazel
@@ -4,12 +4,10 @@ go_library(
     name = "python",
     srcs = [
         "discovery_nopy.go",
-        "discovery_types.go",
         "embed_nopy.go",
         "version_nopy.go",
     ],
     cgo = True,
     importpath = "github.com/DataDog/datadog-agent/pkg/collector/python",
     visibility = ["//visibility:public"],
-    deps = ["//comp/core/autodiscovery/integration"],
 )
